refactor: split ClusteringResult into its own module
config.py was 1382 lines, ClusteringResult plus its disaggregation
helpers and JSON (de)serialization accounted for 871 of them — more
than half the file. ClusteringResult is not a configuration class; it
holds the result of a clustering and the parameters needed to apply
it to new data. Living in config.py muddled the module's purpose and
forced anyone navigating clustering-result code to scroll past every
config dataclass.

Move ClusteringResult and the helpers used only by it
(_expand_periods, _expand_segments_to_timesteps,
_validate_disaggregate_input, _time_index_to_dict /
_time_index_from_dict, _infer_resolution, _get_version) to
src/tsam/clustering_result.py.

The two representation helpers (_representation_to_dict /
_representation_from_dict) stay in config.py because ClusterConfig
and SegmentConfig also use them.

Backwards compatibility for `from tsam.config import ClusteringResult`
is preserved via PEP 562 module __getattr__ — the import lazily
re-resolves into the new module without re-introducing the dep.

Internal importers (tsam/__init__.py, pipeline/__init__.py,
pipeline/types.py, result.py) now import directly from
tsam.clustering_result. One test was updated to import its
_time_index_* helpers from the new module.

Sizes after split:
- config.py:            1382 → 467 lines
- clustering_result.py:  new, 770 lines

All 559 tests pass unchanged.
