+++ src/tsam/__init__.py
@@ -53,9 +53,9 @@ def __getattr__(name: str):
     raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
 
 
+from tsam.clustering_result import ClusteringResult
 from tsam.config import (
     ClusterConfig,
-    ClusteringResult,
     Distribution,
     ExtremeConfig,
     MinMaxMean,
